Clean up module installation code and add stdlib
We have to install modules because we need capitalize() from stdlib and
the and node_group type from WhatsARanjit-node_manager. Clean it up so
it's still really easy to trace the logic of what's going on.

Include a file-sync commit in the module installation so that the PM and
PMR will have the same code, and there won't be pluginsync differences
between the two even before deploying real user code.