Make the peconf parameter to pe_install optional
This is to allow the same task to be used for upgrades, where the
pe.conf file is expected to be in place already on the target system.

Fix an issue relating to extract dir. Need to extract to a dir with a
version number in it to avoid old, leftover dirs from previous versions.