Introduce TODO status check in upgrade
Also fix some compilemaster related bugs.

It looks like there's legitimately a need to check and wait for the
Orchestrator to be fully ready to roll at a point in the plan. Not sure
how to do that yet, but kind of earmarking it by inserting a run_command
for puppet infra status.