Add orchestrator healthcheck step to upgrade
Using the code from plans/upgrade/primary_set.pp. This gives us the
pause/check we need to wait for the orchestrator to be up and running
before resuming the upgrade actions.

Note: may need to further extend healthcheck later if it turns out we
also need to wait for things like pxp-agents running on the
infrastructure nodes to reconnect