Improve retrieve_and_upload function
Make it so that if the target system already has a file of the right
name and size in the target location, we do not re-upload the file. This
will provide the flexibility for users who cannot use the upload
function to pre-stage the files they need in order to run plans that use
retrieve_and_upload.