Make manage_environment_groups configur param bool
The task doesn't deal with Booleans natively right now, but it will pass
a boolean through as the string "true" or "false" transparently. In
order to present the correct UX to the end user when writing plan
parameter files, make the plan take a Boolean now.

We'll fix the task to natively use Booleans when we change to Bolt apply
for this, instead of using the task.