Clarify manage_environment_node_groups param type
The user experience should be, pass a boolean. Unfortunately there's
some ugliness in the shell script due to the boolean being passed via
environment variable and coming out the other side as a string ("true"
or "false"). For now, until we switch to Bolt apply, we're going to just
sequester this ugliness in the shell script and keep the user experience
a Boolean type.