Fix possible race condition in install_module
The pe_xl::install_module function previously used the code-manager-cli
task to perform a 'file-sync commit' action. However, at the end of this
action the live-dir is not guaranteed to be synchronized yet. This means
if a Puppet run starts fast enough it could happen before any code in
the staging-dir is actually available for use.

This commit changes the code-manager-cli action to be 'commit', which is
a combination of 'file-sync commit' and 'file-sync force-sync', at the
end of which code is guaranteed to be ready for use.