Make token-file optional parameter
Due to it being more difficult to determine a correct default value for
token-file in Puppet code than in the task itself, make the token_file
argument in Puppet optional, and if not passed determine the default
value in Bash code in the task.

Make variable assignment more uniform stylistically.