Fix code_manager task arg reference
- The code_manager task's deploy action was failing because the
environment arg wasn't getting passed in for some reason. Instead of
using the arg I read the environment value from the task's
"PT_environment" var that should be set on the task. We could probably
change it so that action is read from "PT_action" instead of an arg too
if we think this is a better approach. I'm also open to just trying to
fix the reference to the argument and not use environment variables.