(SLV-366) variable name adjustments, compactness
This commit makes some UX changes to varible names to new "target"
variables in use, now that there are several of them. The changes aim to
make each "target" clearly relate to the original input variable it was
based on.

This commit also tries to smooth over the epp template quoted string
issue. It makes the puppetdb_database_host parameter to the template
optional, so that undef is permissable to pass.