Remove "environment" code_manager task param
It was unused, and wouldn't have worked even if somebody did pass a
value to it.

Updated the code_manager.json file to match actual usage. Specifically,
the action param was listed as type enum, but in the configure plan we
were already using a variable in invoking it such that the action didn't
match any of the enum values. So now it's just a string.

Fixes #19