Refactor for modern practices
Initial refactor of pe_xl::unit::install only.

Make use of better input typing. Use get_target and get_targets in order
to operate explicitly on Target objects. This will be useful later in
configure and upgrade plans, for detecting automatically if a target is
using the local transport.

Various cleanup and compacting of code, taking advantage of the above
changes.

Removal of explicit OID numbers from csr_extension_requests.yaml. Bug
was fixed in Puppet such that this is no longer necessary.