Rename module to puppetlabs-peadm
Puppet Enterprise Administration module. This name change is being
driven by a desire to standardize on using this module and approach for
all deployments of PE. The module already supports more architectures
than just Extra Large, so a name change is probably a good idea for that
reason too.

Really, the purpose of this module is to capture and automate lifecycle
administration actions relating to the management of PE. Thus, "peadm".