Remove use of pp_role
The pp_role trusted fact is the most-likely-to-be-used trusted fact we
have. It's not a good idea to require the use of this one for peadm
configuration to work.

Instead, encode the information we need for component identification
into pp_application. Use puppet/ as a prefix, and attach whatever
component role a node has after that.

There's a good argument that we shouldn't use pp_application either.
We're already using it though so we don't lose anything this way, and we
gain the benefit of not using pp_role anymore. So shipping this as an
incremental improvement, with room for more improvement in the future.