Ensure the puppet agent is running post-provision
Previously, the agent was not started on compilers and the replica.
Additionally, because the agent was running on the master and postgresql
nodes during install, that could sometimes interfere with directed runs.

After this commit the Puppet agent is ensured stopped on all systems
during provision to ensure full control of Puppet activity, and the
agent is resumed at the end of configuration.