Modernize upgrade plan
This commit ensures the upgrade plan is functional for the peadm module
over ssh. The upgrade plan has not been well supported in the past.
This commit is a first step and its objective is only base
functionality. Additional features such as input validation and
orchestrator transport support will need to be added iteratively.

As an improvement over the alpha version of this plan, upgrade the
PostgreSQL nodes *first*, so that upgrade of the master/replica can
proceed without error in the event a major PostgreSQL version is
released.

Upgrade plan is also annotated with stage separators to make it easier
to follow what is going on.