Set the certname for PE installer hosts
The only way to set the certname right now for hosts configured with the
PE installer is to pre-stage a puppet.conf file with the certname value
set. This commit does that for PE installer hosts in order to allow all
infrastructure nodes to have their certnames set for them, based on the
Bolt target name.

If the target name is different from the IP address
used to reach the host, then an inventory file should be used.