Rework upgrade sequence to support Standard arch
Previously, when upgrading a standard architecture deployment, after the
master runs the puppet-enterprise-installer script auth for compilers
needs to be restored by a puppet agent run. However, the upgrade plan
only performed a run on the dedicated database host, which doesn't exist
in a standard install (the databases are all on the master).

This commit reworks the upgrade sequence such that the master and
replica recieve puppet runs at the appropriate points, along with the
database targets, to ensure that regardless of which node the
database(s) are located on the node will have a Puppet run to finalize
its config before upgrade moves on to the compilers.

Fixes #73