Reduce several check tasks down to one
This commit consolidates two tasks which check for a service to be
available into one, and provides a convenience function wrapper for
using that task so that the high-level plans are easier to read.

Right now every use of the check just waits until the service is
running, so the implementation puts the wait logic in the task itself.
In the event a future need to actually return true/false if the service
is running or not, it will be easy to refactor into a single
"service_status" task instead, and put the wait logic into the
convenience function wrapper for today's "wait until" uses.