Re-implement ad6ae240 without behavior change
This commit re-implements ad6ae240, but without a discovered behavior
change, where the task would exit with code 1 if the target file to read
did not exist.

The expected and original behavior is that if the file does not exist,
the content key is set to null and the error key is set to a string
value, but the task return code is still 0.