Add direct download option for PE installers
This commit adds a new download_mode parameter to the peadm::provision
and peadm::upgrade plans. The default value for the parameter is
bolthost, which causes PE installation content to be downloaded to the
stagingdir of the workstation running bolt and then uploaded to the
PE hosts. Setting download_mode to direct causes PE installation content
to be downloaded directly to the PE hosts. The direct mode skips uploads,
which offers a significant speed boost when the PE hosts have a route to
the internet and the bolt host is connected by a slow link, such as VPN.