Add util plan to clean up pe.conf on PG nodes
If a PG node has the wrong entry for puppet_master_host or
database_host, then running puppet-enterprise-installer won't upgrade
PostgreSQL as needed. The pe.conf on PG nodes is effectively inert most
of the time, so if it gets out of date there's nothing to fix it.

This optional util plan will ensure the pe.conf has the minimum required
correct values.

This util plan may be integrated into the upgrade plan at a later date.