Eliminate convert configure_node_groups param
Previously, the user would need to specify the configure_node_groups
parameter when running the convert plan. In some circumstances failing
to pass the right value would break the convert operation.

This commit adds computed default values for this parameter so that the
user does not need to pass it.