Do not pin the replica to the PE HA Replica group
The `puppet infra provision` command runs the agent on the replica
before correctly configuring other systems. So, we can't have the
replica be considered a replica during that run.

The only reason it was pre-pinned originally was to make sure PostgreSQL
on other systems configured a pg_ident rule for it, but that's being
handled now at install time by pe.conf variables.