Catch puppetdb sync-wait exit code and continue
When the replica is upgraded and PuppetDB sync doesn't complete within
15 minutes, the upgrade command exits with exitstatus 11. This should
not be a condition in which the upgrade plan fails; the upgrade plan
should proceed forward in this case. There are many reasons why
PuppetDB sync might not complete quickly, but it's not a reason not
to continue upgrading the compilers. Upgrading the remaining
compilers will result in a more stable deployment than not upgrading
them, whatever the reason for PuppetDB sync being incomplete/slow.