Remove tty setting from config example
There is a Bolt bug where if tty is true, apply blocks won't work in
plans. That will get fixed but in the meantime, removing this from the
doc example just in case in the time between now and the bug being fixed
somebody copy-pastes the config into something. We don't want them to
get tripped up by something like that.