Incorporate safety wait into sign_csr
When signing CSRs, it is common that the CSR will have been submitted
just prior. Previously, plans which did both would have to pause after
submitting the CSR to ensure there was time for the server to have fully
processed it before running the sign task, in order to avoid any chance
of ready-to-sign timing problems.

This commit makes the sign_csr task itself attempt to sign up to six
times, waiting one second between attempts, in order to avoid needing to
do that in the plan.