Permit target certnames to be specified by var
This allows Target objects to indicate the certname that should be used
for them by setting a var. Using a var avoids needing to construct new
Target objects when/if the target object doesn't have a name, but the
uri isn't a valid certname. In that circumstance, just set_var(certname,
...) on the target.