Remove bolt-project.yaml
Projects do not support functions, so having this file around is
problematic because it prevents `bolt plan run peadm::*` plans from
executing correctly when run from inside the module folder.

The project.yaml file was originally added to make testing/using the
module easier. However, since then, this has become an option:

    bolt project init my_org --modules puppetlabs-peadm

Given an easy way of downloading and using the module, removal of the
bolt-project.yaml file to fix other problems is the thing to do.