Rename peadm::provision => peadm::install
This is a more intuitive name for what the plan does. "Provision" is
associated in cloud terminology with the creation of new cloud
resources, which this plan does not do.

Historical note: the only reason the plan was called provision to begin
with was out of an attempt not to have it share the same word as its
first subplan, given that install and configure are broken into two
separate subplans.