Rename util/add_cert_ext... => modify_cert_ext...
This commit moves the certificate extension modifying plan to the top
level, then parallelizes it, putting the pieces not intended to be
directly invoked by users in the plans/subplans directory.

Reason for putting in the top-level: changing certificate extensions is
an independently useful utility to have available.

Reason for parallelizing: speed up convert.