Use `allow_any_<action>` syntax to simplify specs
We want our spec tests to focus on testing logic, not just enumerating
which plans/tasks/commands are called. A big value out of these tests is
just making sure they run to completion logically, assuming each of the
tasks/plans doesn't fail.

This commit simplifies the spec tests to allow execution actions by
default, to achieve the value of making sure the plans run to completion
with as little cognative overhead or unnecessary repetition as possible.