Fix GitHub README.md problem
The .github/README.md takes precedence over README.md as being the file
which is displayed on a repo's home page, if present. This commit moves
the Github Actions readme file deeper into the .github directory to
prevent it from displaying the repo's README.md file as the
authoritative content to display on the repo's home page.

Also perform some maintenacne updates on the Github Actions readme file.