Add get_cluster_roles task
This task can introspect an existing cluster (when run on a primary
node) and return information about how the cluster is currently
configured. This will aid in running plans that need to modify the
cluster, by reducing the amount of information the user must supply.

The task returns data about the cluster in both primary/replica form, as
well as A/B form, since different forms are useful in different
scenarios.