Fix mis-named group
PEAdm creates some of its own groups, but also adjusts variables or
parameters on some existing groups. We need to use the correct names of
the existing groups, especially where the DR feature is concerned; we
don't get to rename them. So, even though "DR" is more correct these
days, we need to modify the "HA" group