Fix private designation syntax in private plans
Many plans in this module are not intended to be run directly, and are
instead subplans or utility plans for internal use. This commit updates
the doc comments in the plans to indicate to Bolt that they should not
be shown or displayed when running commands like `bolt plan show`.