Use Target object as key in cert_exts hash
Previously, the $target.name was used as the hash key. This was
originally done to work around a bug in an older version of Bolt; it is
not necessary to use the .name string as the hash key.

Using .name makes it more difficult to deal with when there are no
actual Target objects in an array. So let's stop doing that.