(SUP-2643) Add option to override version assertion
Prior to this commit, there was no way to force an override to the PE
version assertion. This commit adds a new parameter that will override
the PE version check and output a warning when enabled. This is useful
in testing new PE versions that are not yet supported by the module.