Use string keys for hash, instead of Target
When you use Target-type keys in a hash, it breaks the ability of the
plan to call an apply() block in Bolt <= 3.17.0. Not sure if this will
ever be fixed. Using the peadm::certname function on a target as a key
isn't perfect, because we don't have a hard guarantee that all hosts
have unique certnames, but it works better than Target for now. And if
hosts don't have unique certnames, lots more will break besides this.