Make peadm::uninstall private, multi-target
There isn't a reason to restrict this to operating on only one target at
a time. Additionally, we should use the convention $targets, so that the
CLI flag --targets or -t will auto-populate this parameter.

Because this may not universally work on all systems, and we don't
handle the situation where Puppet might be installed, but the
puppet-enterprise-uninstaller script isn't prresent, mark this plan
experimental/private.