Try to ensure LANG is valid
If Ruby starts without a LANG env var, it will assume locale "C". This
is a non-UTF-8 locale, and usually not the system default. If commands
print UTF-8 characters, Ruby will fail to correctly parse them if it
wasn't started with a valid locale.

The main scenario we want to account for is when a privilege escalation
command such as sudo or powerbroker has elevated a shell, but stripped
all environment variables out of it. If that happened, we want to try
and set the default system locale, which should be in
/etc/default/locale.