Plan for adding an external database component
Includes a collection of utility and subplans that can be assembled to
add a new external database componenet to a PEADM deployment, a required
component of the Extra Large architecture. Databases can be added in
both DR and no-DR deployments or if a database is lost, replaced using
a still functioning replica as a source of data.