(SOLARCH-581) Refactor backup plan
This commit refactors the peadm::backup plan for clarity. A commit to
refactor the peadm::restore plan will follow.

Note that for now, it is impossible to backup the CA. This is because we
need to invest more time making that process work correctly. At the
moment, it mostly destroys servers.