Ensure using correct primary
When reusing failed infrastructure components they may be configured for
a different primary then is current and have an old certificate
revacation list. Commit ensures that agent configuration is updated for
the current primary and fetches CRL from that primary.

Includes a little cleanup lifted from the add_compiler plan.