Wrap some add_replica actions in an if statement
Actions taken on $replica_postgresql_host were previously a noop when
the parameter was not provided because the tasks would be ran with an
undef set of targets. Doing this when introducing get_psql_version
results in an error because you can't run functions against an undef
return value if the task was a noop.

Wrap everything related to $replica_postgresql_host in an if statement
to prevent errors when the parameter is not provided.