Add descriptive error messages to fail_on_transport
This commit updates the `fail_on_transport` function
with an optional `message` parameter that can be
used to supply an error message.

The `peadm::upgrade` and `peadm::subplans::modify_certificate`
plans are also updated to use this parameter to provide
an explanation of why use of the "pcp" transport is resulting
in a failure along with suggestions for how to configure
the plans for success.