Fail peadm::util::retrieve_and_upload on PCP
The `peadm::util::retrieve_and_upload` plan downloads a
PE installer tarball and then uploads it to a set of
hosts that need it (Primary and Database nodes).

When the PCP transport is in use, this upload is done
via the `bolt_shim::upload` task, which essentially
reads the file into memory, encodes it, and then
ships it over Orchestrator as a task argument.

This method of file transfer is utterly inadequate
for shipping files as large as the PE installer
which is over half of a gigabyte in size. Use
of "bolt_shim::upload" will cause either a
Linux kernel OOM, an Orchestrator OOM, or
slam into the `max-message-size` allowed by
the Orchestrator.

This commit updates `peadm::util::retrieve_and_upload`
to fail with a descriptive error message if any
of the target nodes is using the PCP transport.