Adding /etc/puppetlabs/enterprise/conf.d/pe.conf
When a database is added outside of the peadm::install plan, the /etc/puppetlabs/enterprise/conf.d/ directory structure is not created. When the peadm::upgrade plan is run, it fails because it can't write the pe.conf file because the directory doesn't exist. The contents of the file aren't that important, they get rewritten when the upgrade runs. This change just ensures the directory is there for it.