change default pe tarball paths
Changes stagingdir default to /var/tmp and adds uploaddir param for the same
purpose on targets. Using /tmp for very large files can be problematic as it
is memory backed and typically not limited by default. With ~1.4G used by
the PE installer, unexpected bolt failures on systems with low to no swap
occur (high performance or containers).