peadm::install: Depend code-manager setup on r10k remote presence, not r10k ssh key
We've logic in place that checks if the user provided an r10k remote key
and then automatically enable code-manager. This doesn't make sense
because the key is optional. We need to depend on a r10k URL. The URL
can be an absolute path to a local repo or an https url, both don't
require a remote key.

Fixes a bug I introcued in https://github.com/puppetlabs/puppetlabs-peadm/pull/352/files#r1373941616