Adding support for legacy compilers (#448)
* Adding support for legacy compilers

* Adding inventoryfile to test upgrade legacy workflow

* feat(documentation): add instructions for converting compilers to legacy

This commit introduces a new section in the `convert.md` documentation. It provides instructions on how to convert compilers to legacy compilers for Puppet Enterprise installations using puppetlabs-peadm version 3.21 or later, as well as for versions prior to 3.21. The new section includes specific commands to run and references to other relevant documentation.

* Fixing lint and regenerating referencemd

* PE-38772 Node groups added for legacy compilers (#455)

* Fixing typo for parameter in docs

---------

Co-authored-by: Ioannis Karasavvaidis <ioannis@puppet.com>
Co-authored-by: Neil Anderson <neil.anderson@perforce.com>
Co-authored-by: Aaron Shannon <aaron.shannon@puppet.com>