(PE-38770) Install Plan accepts legacy_compilers key (#474)
* feat(peadm): add support for legacy compilers

- Added `legacy_compilers` parameter to `peadm::install`, `peadm::subplans::install`, and `peadm::subplans::configure` plans.
- Updated `peadm::assert_supported_architecture` function to handle `legacy_compilers`.
- Modified various plans to include `legacy_compilers` in the installation and configuration processes.
- Updated documentation to reflect the changes and added examples for `legacy_compilers`.

This change allows the PEADM module to support legacy compilers in addition to the standard compilers.

* feat(assert_supported_architecture): normalize and combine compiler variables

- Normalize `$legacy_compilers` and `$compiler_hosts` to arrays, handling both strings and arrays.
- Combine the normalized arrays into `$all_compilers`.
- Set `$has_compilers` to `undef` if `$all_compilers` is empty, otherwise set it to `true`.
- Update tests to cover scenarios with both `compiler_hosts` and `legacy_compilers`.

This change ensures that the function correctly handles different types for compiler variables and improves robustness.

* Updated REFERENCE.md to document the new parameter.

* style(tests): use single quotes for consistency in spec files

Updated all double quotes to single quotes in the `assert_supported_architecture_spec.rb` and `install_spec.rb` files to maintain consistency in string literals across the test suite.

* refactor(plans): rename legacy_targets to legacy_compiler_targets

Renamed the variable `legacy_targets` to `legacy_compiler_targets` across multiple plans for better clarity and consistency. This change affects the following plans:
- convert_compiler_to_legacy.pp
- subplans/configure.pp
- subplans/install.pp

The new variable name more accurately reflects its purpose, which is to target legacy compiler hosts.

* docs(install): correct grammar in installation requirements section

Corrected the grammar in the installation requirements section by changing "Puppets must not be installed" to "Puppet must not be installed" for clarity and accuracy.