(#469) Upgrade: Assign production environment to node groups
This makes the environment for node groups configureable. The default is
'production'. This is also the default in the node_group resource type:

https://github.com/puppetlabs/puppetlabs-node_manager/blob/main/lib/puppet/type/node_group.rb#L124

We could also default to `undef` in PEADM and rely on the node_manage
module, but explicitly setting `production` seems to be safer.

This is related to
https://github.com/puppetlabs/puppetlabs-peadm/issues/469 and makes the
environment configureable during upgrades.

Co-authored-by: David Mallon <david.mallon@perforce.com>