(PE-39352,PE-39351) Gather PE version in get_peadm_config task
Adds a lookup of the pe_build file to provide the PE version in the hash
returned by the get_peadm_config task.

We will use this in add_database and backup/restore plans as a version
gate for handling the newer host_action_collector database that is only
present in PE 2023.7+, for example.