(PE-39352) Update backup/restore plans for hac database
PE 2023.7.0 added the host-action-collector service and the
backup/restore plans need to be made aware of the pe-hac database.

This patch adds a version test, best on the addition of pe_version check
of /opt/puppetlabs/server/pe_build in the peadm::get_peadm_config task.

If >= 2023.7, then pe-hac is added to the defaults returned by the
various recover_opts functions that define which resource backup/restore
should wrangle.

The patch does not attempt to version test the user's custom override
hash of selections, on the assumption that they know, or should be
allowed to specify whatever is needed.

If pe_version is not found (damaged cluster, missing from
peadm_config.json), the patch defaults to assuming earliest version,
which will skip pe-hac. This favors the plans running without error at
the cost of possibly missing pe-hac in a newer installation.
Alternately, we could reverse this and incur failures in damaged older
clusters requiring use of $custom to work around.

Finally there is an edge case it does not address which is the case of
newer pe versions with pe-hac restoring tarballs that were intended to
include all databases, but which were created with peadm predating this
patch. These will fail the pg-restore pe-hac commands, since pe-hac
won't be in the backup, and again require use of $custom to work around.