(PE-39351) Update add_database plan to work with patching/HAC (#531)
* (PE-39351) Update add_database plan to work with patching/HAC

This commit updates the add_database plan to include the pe-patching and
pe-hac databases when it cleans up the original postgres on the primary.

The add_database plan uses pg_basebackup to copy the entire contents of
the postgres DB from the original primary DB to the new DB node's
postgres. That process doesn't individually update a list of databases,
so there's no context of "which DBs to move", because it just moves
everything. However, during the cleanup process on the original DB That
exists on the primary, the plan _does_ need to list the individual
internal postgres dbs that need to be deleted.

* Updating reference.md

* Fixing lint

---------

Co-authored-by: Neil Anderson <neil.anderson@perforce.com>