(#469) Upgrade: Assign correct environment to node groups  (#635)
* (#469) Upgrade: Assign production environment to node groups

This makes the environment for node groups configureable. The default is
'production'. This is also the default in the node_group resource type:

https://github.com/puppetlabs/puppetlabs-node_manager/blob/main/lib/puppet/type/node_group.rb#L124

We could also default to `undef` in PEADM and rely on the node_manage
module, but explicitly setting `production` seems to be safer.

This is related to
https://github.com/puppetlabs/puppetlabs-peadm/issues/469 and makes the
environment configureable during upgrades.

Co-authored-by: David Mallon <david.mallon@perforce.com>

* (#469) Upgrade: Assign correct environment to node groups

This checks if a user configured a environment in pe.conf. If that's the
case, it will be used for the PEADM-specific node groups. Otherwise we
fall back to production.

This fixes a timing issue discovered in #469. In situations where the PE
infra isn't running in production, we cannot assume that a production
environment exists. And a node group can only reference classes from the
environment the node group belongs to.

---------

Co-authored-by: David Mallon <david.mallon@perforce.com>

Note: internal tickets exist for the 2 CI failures, we will merge past them for now